backendservice/requirements.txt

Flask==2.0.3
PyYAML==5.3.1
